카드 컬럼 이동

+++ src/main/java/com/taskrail/service/CardService.java
@@ -3,7 +3,11 @@ package com.taskrail.service;
 import com.taskrail.dto.CardRequestDto;
 import com.taskrail.dto.CardResponseDto;
 import com.taskrail.entity.Card;
+import com.taskrail.entity.Columnss;
+import com.taskrail.entity.User;
 import com.taskrail.repository.CardRepository;
+import com.taskrail.repository.ColumnssRepository;
+import com.taskrail.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -16,36 +20,104 @@ import java.util.List;
 @Slf4j
 public class CardService {
     private final CardRepository cardRepository;
-    private final ColumnRepository columnRepository;
+    private final ColumnssRepository columnssRepository;
+    private final UserRepository userRepository;
     public List<CardResponseDto> getCards(Long columnId) {
         return cardRepository.findAllByColumn_IdOrderByOrdersDesc(columnId).stream().map(CardResponseDto::new).toList();
     }
 
-    public void createCard(Long columnId, CardRequestDto requestDto) {
-        Column column = columnRepository.findById(columnId).orElseThrow(
-                ()->new IllegalArgumentException("컬럼이 없습니다.")
+    public void createCard(Long columnId, CardRequestDto requestDto, User user) {
+        userRepository.findById(user.getId()).orElseThrow(
+                ()->new NullPointerException("유저가 없습니다.")
         );
+        Columnss column = columnssRepository.findById(columnId).orElseThrow(
+                ()->new NullPointerException("컬럼이 없습니다.")
+        );
+
         List<Card> cards = cardRepository.findAll();
 
-        Long orders = cards.get(cards.size() - 1).getOrders();
+        try{
+            Long orders = cards.get(cards.size() - 1).getOrders();
+            Card card = new Card(requestDto, column, orders+1, user);
+            cardRepository.save(card);
+        }catch (IndexOutOfBoundsException e){
+            Long orders = 1L;
+            Card card = new Card(requestDto, column, orders, user);
+            cardRepository.save(card);
+        }
 
-        Card card = new Card(requestDto, column, orders+1);
-        cardRepository.save(card);
+
     }
 
-    public void deleteCard(Long cardId) {
+    public void deleteCard(Long cardId, User user) {
         Card card = cardRepository.findById(cardId).orElseThrow(
-                () -> new IllegalArgumentException("카드가 없습니다.")
+                () -> new NullPointerException("카드가 없습니다.")
         );
+
+        if(!card.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("작성한 유저가 아닙니다.");
+        }
         cardRepository.delete(card);
     }
 
 
     @Transactional
-    public void updateCard(Long cardId, CardRequestDto requestDto) {
+    public void updateCard(Long cardId, CardRequestDto requestDto, User user) {
         Card card = cardRepository.findById(cardId).orElseThrow(
                 () -> new IllegalArgumentException("카드가 없습니다.")
         );
+
+        if(!card.getUser().getId().equals(user.getId())){
+            throw new IllegalArgumentException("작성한 유저가 아닙니다.");
+        }
         card.update(requestDto);
     }
+
+    public CardResponseDto getCardOne(Long cardId) {
+        Card card = cardRepository.findById(cardId).orElseThrow(
+                () -> new IllegalArgumentException("카드가 없습니다.")
+        );
+
+        return new CardResponseDto(card);
+    }
+
+    @Transactional
+    public void updatePrevCard(Long cardId) {
+        Card card = cardRepository.findById(cardId).orElseThrow(
+                () -> new IllegalArgumentException("카드가 없습니다.")
+        );
+
+        if(card.getColumn().getId() == 1){
+            throw new IllegalArgumentException("앞으로 이동할 수 없습니다.");
+        }
+        Columnss column = columnssRepository.findById(card.getColumn().getId()-1).orElseThrow(
+                ()-> new IllegalArgumentException("앞으로 이동할 수 없습니다.")
+        );
+        //jpa sql 직접 사용해보기, 어노테이션 query
+        // update
+        // set columnss = columnss -1
+        // where cardId = cardId
+
+
+        card.updatePrev(column);
+    }
+
+    @Transactional
+    public void updateNextCard(Long cardId) {
+        Card card = cardRepository.findById(cardId).orElseThrow(
+                () -> new IllegalArgumentException("카드가 없습니다.")
+        );
+        List<Card> cards = cardRepository.findAll();
+
+        Long lastIndex = cards.get(cards.size() - 1).getColumn().getId();
+
+        if(card.getColumn().getId().equals(lastIndex)){
+            throw new IllegalArgumentException("뒤로 이동할 수 없습니다.");
+        }
+        Columnss column = columnssRepository.findById(card.getColumn().getId()+1).orElseThrow(
+                ()-> new IllegalArgumentException("뒤로 이동할 수 없습니다.")
+        );
+        card.updateNext(column);
+    }
+
 }
